+++ src/rust/lqosd/src/node_manager/ws/ticker/ipstats_conversion.rs
@@ -46,8 +46,8 @@ impl From<&IpStats> for IpStatsWithPlan {
                 };
                 result.ip_address = format!("{}", name);
                 result.plan = DownUpOrder {
-                    down: circuit.download_max_mbps as f32,
-                    up: circuit.upload_max_mbps as f32,
+                    down: circuit.download_max_mbps,
+                    up: circuit.upload_max_mbps,
                 };
             }
         }
